Blackbox testing for saving goal controller
Implemented some blackbox testing for the endpoints handled by the saving goal controller
